Add node removal blocking/draining knob based on nodes status
This PR ensures OSP nodes are not removed until the node has been
drained (by a drain pod). It handles it through CRD status updates
and by creating a blocker pod per OSP node, which cannot be removed
by OCP due to the usage of finalizers.

For now the drain pod is a sample container that only sleeps for
5 minutes. Next step would be to accept the drain pod image and/or
parameters. Once the drain pod is completed, the finalizer is removed
from the blocker pod, and the node can be deleted by OCP.