Manage selinux using CRD
So far selinux was disable. Make this a parameter
to be managable. We use a machineconfig for this to set a kernel
parameter when SelinuxDisabled: true is set in the CR.

Until worker nodes support working with selinux in enforcing mode
the default in the example CR is disabled.

When performance tuning operator is integrated for cpu pinnig and
other kernel parameters, lets switch to that as well. Therefore
no handling when flipping from true to false is there which means
the machineconfig is still there when we flip from selinux
disabled to enabled.