Run OSP scale down task on node removal
This change introduce running two pods during scale down actions:

First <node-name>-drain-pre runs before node got removed. It
disables the nova-compute service via OSP API call and optionally
triggers live migration if specified in the section of a specific
node to remove. If the migration fails the pod will proceed checking
until all instances got manually removed from the compute nodes.

Note: right now only live migration is handled (ACTIVE instances).
      If wanted a follow up is required to improve the pre script
      to handle instances in off/error state.

Second pod <node-name>-drain-post runs after the node got removed
to cleanup OSP entries via OSP api. It removes the nova-compute
service entry and also removes the removed node as a placement
resource provider.

A new configmap get created in the openshift-machine-api namespace
which provides the scripts to run in each of the step.

The change introduce 5 new CR parameters to get the information
on how to connect to the OSP API:

- openStackClientAdminSecret: Secret which holds the admin user
details.
- openStackClientConfigMap: ConfigMap which holds the general
information on howto connect to the OSP API
- openStackNamespace: Namespace which holds the above Secret/ConfigMap.
If not specified the default 'openstack' is used.
- Drain paremeters used for the pod running the pre/post script.
it requires openstackclient and ocs-placement to be installed.
  drain:
    drainPodImage: url to the image used
    enabled: true/false

enabled: global switch to enable live migration in <node-name>-drain-pre
job. If not set, the default is false. This can be overwritten per node
when NodesToDelete is used.