Add support for selecting the node to remove upon scale down
This PR adds a new option to the CR spec to allow the user to both
select the node to scale down and the draining process. The Spec
field added is:
  spec:
    ...
    nodesToDelete:
    - name: node-name (e.g., worker-3)
      drain: true|false
    - name: ...

Note the drain is optional and considered false if not set. The
functionality behind the drain will come in follow up PRs. The idea
behind is that if drain is set to false, the user will need to drain
the node and the operator will wait for it to delete the node. If
it is set to true, and automatic draining will be triggered (with
no guarantees of finishing).

More than one node can be added to the nodesToDelete list, and what
the operator does is to add the required annotations into their
associated machine objects, so that those are prioritized for removal
once the scale down is triggered. If nothing stated, the operator
will choose the newest machine.

In addition this PR adds two extra things:
- Reconcile every 10 mins, regardless of changes to the CR or the
associated objects being watched
- Ensure if machineset replicas are adjusted, the CR get updated
with that information