Fix problem with loop at reconciling
Seems if infra-pods are not deployed there is a situation where
the updateNode status will be looping forever as node Status was
not being updated and an update of the CR is triggered anyway,
raising another reconcile.

To avoid this situation, we ensure we annotate the status for
NotReady nodes, as well as to ensure the infra-pods are reconciled
even if from the CR perspective that has already happened. This
ensures if they got deleted or modified, they will be
re-applied/updated