Disable nova-compute services early
Right now we only disable the current compute worker node which
is deleted. When multiple compute worker nodes get removed at the
same time this could lead to migrating instances from the current
node to a compute worker node which gets removed next.

With this change we monitor machinev1beta1.Machine{} and map them
to the machineset to get the OSP worker nodes. Checking the
deletionTimestamp get the marked deleted compute worker nodes.
Those are the nodes which get removed. The nova-compute worker nodes
for all of them get disabled all at once. With this we should not
run in a situaton where either manual or automatic draining the compute
nodes migrate instances to a compute nodes which gets removed next.