Disable all deletedTagedNodes in drain pre job
Right now we only disable the current compute worker node which
is deleted. When multiple compute worker nodes get removed at the
same time this could lead to migrating instances from the current
node to a compute worker node which gets removed next.

With this change we pass a list of tagged deleted nodes to the pre
drain job and disable the compute service of not already disabled.
This makes sure that all deleted tagged nova-compute services get
disabled before any draining starts (manual/automatic) after the
scale down operation got triggered.

This is a simplification of previous approach taken in [1].

[1] https://github.com/openstack-k8s-operators/compute-node-operator/pull/15