Add finalizer to cleanup blocker-pods on CR delete
Right now when a compute-node-operator CR gets deleted without
scaling down worker nodes to 0 the finalizer on the blocker-pods
don't get removed and while the machine objects get deleted, the
node objects are stuck in SchedulingDisabled until the finalizer
on the blocker pods get manually removed.

This introduces a finalizer on the compute-node-operator CR to
trigger cleanup of the remaining compute nodes.