Add option to have either dedicated or shared compute workers
The new option on the CRD is named "dedicated". By default is False
if not specified. If set to false, the taint is not set on the
machineset and therefore not added to the generated worker nodes
either. That will allow any OpenShift pod to run on OpenStack worker
nodes. By contrast, if the node is dedicated, that means the taint is
added and therefore only infra pods can be executed there, like the
ones needed for monitorization, OpenShift SDN, machine configuration,
etc.